Refactor status messages
The buildresult passed as a parameter is in capitals. However, I would not like our slack messages to contain 'shouting' statuses. Also, there is a need to add an 'Unknown' status for any other non-covered case of result.

Thus, I created a new 'status' variable that takes its value according to the buildresult and is displayed in the slack message.